fix: Fix template type mismatch error for formatNumber function (v0.53.15)
The formatNumber template function expects int64 parameters, but FilesData,
HardlinksData, and ScansData struct Total fields were defined as int. This
caused template execution errors when rendering pages with file counts.

Changed Total field type to int64 in all affected structs and added
appropriate type conversions when assigning database query results.

+++ internal/server/handlers.go
@@ -402,7 +402,7 @@ func (s *Server) HandleFiles(w http.ResponseWriter, r *http.Request) {
 
 	data := FilesData{
 		Files:             filesWithUsage,
-		Total:             total,
+		Total:             int64(total),
 		Page:              int64(page),
 		Limit:             limit,
 		TotalPages:        CalculateTotalPages(total, limit),
@@ -584,7 +584,7 @@ func (s *Server) HandleHardlinks(w http.ResponseWriter, r *http.Request) {
 
 	data := HardlinksData{
 		Groups:     groups,
-		Total:      total,
+		Total:      int64(total),
 		Showing:    len(groups),
 		Page:       int64(page),
 		TotalPages: CalculateTotalPages(total, limit),
@@ -642,7 +642,7 @@ func (s *Server) HandleScans(w http.ResponseWriter, r *http.Request) {
 
 	data := ScansData{
 		Scans:      scanDisplays,
-		Total:      total,
+		Total:      int64(total),
 		Page:       int64(page),
 		TotalPages: CalculateTotalPages(total, limit),
 		Title:      "Scan History",
